refactor code
- change the theme of the navigation bar to use that defined by the apps
  theme.
- Change AdminSectionNavItem from a statfule to statless widget
- Add a Getx controller to manage teh state of AdminSectionNavItem
- Moved all the code for the navigation bar into the navigation_bar
  folder in the widgets directory.
